+++ components/Courses/RelatedCoursesSection.tsx
@@ -27,7 +27,7 @@ const RelatedCoursesSection = ({ courses }: RelatedCoursesSectionProps) => {
         </button>
       </div>
 
-      <div className="grid grid-cols-2 gap-4 p-6 md:grid-cols-3 lg:grid-cols-4">
+      <div className="grid grid-cols-1 gap-4 p-6 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4">
         {courses.map((course, index) => (
           <CourseCard
             key={index}
